+++ TileObject.cs
@@ -26,51 +26,4 @@ namespace EpicGameEngine
     }
 
 
-    public class ImmovableObject : TileObject
-    {
-        public ImmovableObject(Tile tile) : base(tile)
-        {
-        }
-
-        public override void OnCollision(TileObject otherTileObj)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public class MovableObject : TileObject
-    {
-        public MovableObject(Tile tile) : base(tile)
-        {
-        }
-
-        public override void OnCollision(TileObject otherTileObj)
-        {
-            if (this.Tile.Position.x == otherTileObj.Tile.Position.x && this.Tile.Position.y == otherTileObj.Tile.Position.y)
-            {
-                if (otherTileObj.GetType() == MovableObject)
-                {
-                    // Make an interaction between it and other tile object
-                }
-                else if (otherTileObj.GetType() == ImmovableObject)
-                {
-                    // Make an interaction based on this
-                }
-            }
-        }
-        public virtual void MoveTo(Tile tile)
-        {
-            this.Tile = tile;
-            this.Tile.TileObject = this; 
-            // Make TileObject move to destination Tile
-        }
-        public virtual bool IsTileAValidDest(Tile tile)
-        {
-            if (tile.IsWalkable == true)
-            {
-                return true;
-            }
-            return false;
-        }
-    }
 }
